Add files via upload
I designed a web-based GUI for people who want to use the Marauder software like me — without a screen and without a battery. It may have some shortcomings and is open to further development. It’s very easy to use: connect the Marauder to your computer via USB, open the HTML file, then connect to the device and continue. Enjoy life 😎